+++ apps/backend/clixon_backend_handle.c
@@ -86,6 +86,8 @@ struct backend_handle {
     int                      bh_magic;     /* magic (HDR)*/
     clicon_hash_t           *bh_copt;      /* clicon option list (HDR) */
     clicon_hash_t           *bh_data;      /* internal clicon data (HDR) */
+    event_stream_t          *ch_stream;   /* notification streams, see clixon_stream.[ch] */
+    
     /* ------ end of common handle ------ */
     struct client_entry     *bh_ce_list;   /* The client list */
     int                      bh_ce_nr;     /* Number of clients, just increment */