+++ lib/src/clixon_handle.c
@@ -53,6 +53,8 @@
 #include "clixon_log.h"
 #include "clixon_err.h"
 #include "clixon_yang.h"
+#include "clixon_xml.h"
+#include "clixon_stream.h"
 #include "clixon_options.h"
 
 #define CLICON_MAGIC 0x99aafabe
@@ -65,9 +67,10 @@
  * @see struct backend_handle
  */
 struct clicon_handle {
-    int                ch_magic;    /* magic (HDR) */
-    clicon_hash_t     *ch_copt;     /* clicon option list (HDR) */
-    clicon_hash_t     *ch_data;     /* internal clicon data (HDR) */
+    int               ch_magic;    /* magic (HDR) */
+    clicon_hash_t    *ch_copt;     /* clicon option list (HDR) */
+    clicon_hash_t    *ch_data;     /* internal clicon data (HDR) Unclear why two? */
+    event_stream_t   *ch_stream;   /* notification streams, see clixon_stream.[ch] */
 };
 
 /*! Internal call to allocate a CLICON handle. 
@@ -133,6 +136,7 @@ clicon_handle_exit(clicon_handle h)
 	hash_free(copt);
     if ((data = clicon_data(h)) != NULL)
 	hash_free(data);
+    stream_free(clicon_stream(h));
     free(ch);
     return 0;
 }
@@ -171,4 +175,27 @@ clicon_data(clicon_handle h)
     struct clicon_handle *ch = handle(h);
 
     return ch->ch_data;
+}
+
+/*! Return stream hash-array given a clicon handle.
+ * @param[in]  h        Clicon handle
+ */
+event_stream_t *
+clicon_stream(clicon_handle h)
+{
+    struct clicon_handle *ch = handle(h);
+
+    return ch->ch_stream;
+}
+
+int
+clicon_stream_append(clicon_handle h,
+		     event_stream_t *es)
+{
+    struct clicon_handle *ch = handle(h);
+    event_stream_t      **ep;
+
+    for (ep = &ch->ch_stream; (*ep); ep=&(*ep)->es_next);
+    *ep = es;
+    return 0;
 }